+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 # Changelog
 
-## [0.2.0]
+## [0.2.2]
 ### Bug
+- Options are not saved when updating a filter using string.
+
+## [0.2.1]
+### Bug
+- Addressing issues with single and double quote characters inside filtered values.
+
+## [0.2.0]
+### Added
 - Adding ability to compare with variables.
 - Adding condition priority.